% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bits_LShift}
\alias{bits_LShift}
\title{Shift Bits to the LEFT}
\usage{
bits_LShift(arr, b)
}
\arguments{
\item{b}{int to shift}

\item{a}{INTEGER to be manipulated}
}
\value{
Updated INTEGER a appropriately shifted
}
\description{
Shift Bits to the LEFT
}
