% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bits_RShift}
\alias{bits_RShift}
\title{Shift Bits to the Right}
\usage{
bits_RShift(arr, b)
}
\arguments{
\item{b}{int to shift}

\item{a}{INTEGER to be manipulated}
}
\value{
Updated INTEGER a appropriately shifted
}
\description{
Shift Bits to the Right
}
