% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_AND}
\alias{cpp_AND}
\title{Pairwise AND operations}
\usage{
cpp_AND(arr, brr)
}
\arguments{
\item{a}{INTEGER to be manipulated}

\item{b}{INTEGER to be manipulated}
}
\value{
Updated INTEGER after AND
}
\description{
Pairwise AND operations
}
