% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_base2int}
\alias{cpp_base2int}
\title{Convert an string of base back to BASE10 INTEGER}
\usage{
cpp_base2int(str, base = 16L)
}
\arguments{
\item{base}{INTEGER of base (2,32)}

\item{num}{INTEGER to be converted}
}
\value{
string of integer in the new base
}
\description{
Convert an string of base back to BASE10 INTEGER
}
