% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_btrim}
\alias{cpp_btrim}
\title{String Both Trim}
\usage{
cpp_btrim(str, t = " \\t\\n\\r\\f\\v")
}
\arguments{
\item{t}{Trimming elements}

\item{s}{String to be trimmed}
}
\value{
Updated String s, trimmed (both left and right)
}
\description{
String Both Trim
}
