% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_explode}
\alias{cpp_explode}
\title{Explode string into array based on separator}
\usage{
cpp_explode(sep, str)
}
\arguments{
\item{sep}{String to determine how split}

\item{str}{String to be exploded (split)}
}
\value{
vector of strings (List array as CharacterVector)
}
\description{
Explode string into array based on separator
}
