% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gcd}
\alias{cpp_gcd}
\title{Compute gcd	(Euler Method)}
\usage{
cpp_gcd(x, y)
}
\arguments{
\item{x}{first integer}

\item{y}{second integer}
}
\value{
INTEGER (gcd)
}
\description{
Compute gcd	(Euler Method)
}
