% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gcd_lcm}
\alias{cpp_gcd_lcm}
\title{Compute gcd / lcm simultaneously (Euler Method)}
\usage{
cpp_gcd_lcm(x, y)
}
\arguments{
\item{x}{first integer}

\item{y}{second integer}
}
\value{
two INTEGER values in list form (gcd) and (lcd)
}
\description{
Compute gcd / lcm simultaneously (Euler Method)
}
