% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_implode}
\alias{cpp_implode}
\title{Implode string array into string based on separator}
\usage{
cpp_implode(r, sep = "")
}
\arguments{
\item{sep}{String to determine how join}

\item{vector}{of strings (array as CharacterVector)}
}
\value{
joined String
}
\description{
Implode string array into string based on separator
}
