% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_int2base}
\alias{cpp_int2base}
\title{Convert an integer to a string of base}
\usage{
cpp_int2base(num, base = 16L)
}
\arguments{
\item{num}{INTEGER to be converted}

\item{base}{INTEGER of base (2,32)}
}
\value{
string of integer in the new base
}
\description{
Convert an integer to a string of base
}
