% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_lcm}
\alias{cpp_lcm}
\title{Compute lcd	(Euler Method)}
\usage{
cpp_lcm(x, y)
}
\arguments{
\item{x}{first integer}

\item{y}{second integer}
}
\value{
INTEGER (lcd)
}
\description{
Compute lcd	(Euler Method)
}
