% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_md5}
\alias{cpp_md5}
\title{Compute md5 of CharacterVector}
\usage{
cpp_md5(str, times = 1L)
}
\arguments{
\item{str}{CharacterVector of Strings}

\item{times}{INTEGER on how many iterations on single string}
}
\value{
CharacterVector of MD5(times)
}
\description{
Compute md5 of CharacterVector
}
