% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_OR}
\alias{cpp_OR}
\title{Pairwise OR operations}
\usage{
cpp_OR(arr, brr)
}
\arguments{
\item{a}{INTEGER to be manipulated}

\item{b}{INTEGER to be manipulated}
}
\value{
Updated INTEGER after OR
}
\description{
Pairwise OR operations
}
