% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_primes}
\alias{cpp_primes}
\title{Compute NumericVector of Primes}
\usage{
cpp_primes(n, first = FALSE)
}
\arguments{
\item{n}{How Many}

\item{first}{BOOLEAN (if FALSE) returns primes <= 'n' ELSE returns 'n' primes}
}
\value{
NumberVector of PRIMES
}
\description{
(ported from pracma::primes) with new 'n' primes functionality
SLOWS above 10^7
}
