% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_RShift}
\alias{cpp_RShift}
\title{Shift Bits to the Right}
\usage{
cpp_RShift(a, b)
}
\arguments{
\item{a}{INTEGER to be manipulated}

\item{b}{int to shift}
}
\value{
Updated INTEGER a appropriately shifted
}
\description{
Shift Bits to the Right
}
