% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_SHIFT_R}
\alias{cpp_SHIFT_R}
\title{Shift Bits to the Right}
\usage{
cpp_SHIFT_R(arr, b)
}
\arguments{
\item{b}{int to shift}

\item{a}{INTEGER to be manipulated}
}
\value{
Updated INTEGER a appropriately shifted
}
\description{
Shift Bits to the Right
}
