% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_str_replace}
\alias{cpp_str_replace}
\title{Search/Replace a String Subject}
\usage{
cpp_str_replace(search, replace, subject)
}
\arguments{
\item{String}{'subject'}
}
\value{
updated 'subject' String appropriate replaced ... (no REGEX here)
}
\description{
Search/Replace a String Subject
}
