% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_str_reverse}
\alias{cpp_str_reverse}
\title{Reverse a String s}
\usage{
cpp_str_reverse(str)
}
\arguments{
\item{String}{to 'repeat'}
}
\value{
updated String reversed
}
\description{
Reverse a String s
}
