% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_strtolower}
\alias{cpp_strtolower}
\title{String to lower case}
\usage{
cpp_strtolower(str, locale = "en_US.UTF-8")
}
\arguments{
\item{s}{String to be transformed}
}
\value{
Updated String s, lower cased
}
\description{
String to lower case
}
