% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_strtoupper}
\alias{cpp_strtoupper}
\title{String to upper case}
\usage{
cpp_strtoupper(str, locale = "en_US.UTF-8")
}
\arguments{
\item{s}{String to be transformed}
}
\value{
Updated String s, upper cased
}
\description{
String to upper case
}
