% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_tolower}
\alias{cpp_tolower}
\title{String to lower case}
\usage{
cpp_tolower(s)
}
\arguments{
\item{s}{String to be transformed}
}
\value{
Updated String s, lower cased
}
\description{
String to lower case
}
