% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_toupper}
\alias{cpp_toupper}
\title{String to upper case}
\usage{
cpp_toupper(s)
}
\arguments{
\item{s}{String to be transformed}
}
\value{
Updated String s, upper cased
}
\description{
String to upper case
}
