% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_trim}
\alias{cpp_trim}
\title{String Trim}
\usage{
cpp_trim(s, t = " \\t\\n\\r\\f\\v")
}
\arguments{
\item{s}{String to be trimmed}

\item{t}{Trimming elements}
}
\value{
Updated String s, trimmed (both left and right)
}
\description{
String Trim
}
