% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_XOR}
\alias{cpp_XOR}
\title{Pairwise XOR operations}
\usage{
cpp_XOR(arr, brr)
}
\arguments{
\item{a}{INTEGER to be manipulated}

\item{b}{INTEGER to be manipulated}
}
\value{
Updated INTEGER after XOR
}
\description{
Pairwise XOR operations
}
