% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s_implode}
\alias{s_implode}
\title{Implode string array into string based on separator}
\usage{
s_implode(sep, r)
}
\arguments{
\item{sep}{String to determine how join}

\item{vector}{of strings (array as CharacterVector)}
}
\value{
joined String
}
\description{
Implode string array into string based on separator
}
