% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s_ltrim}
\alias{s_ltrim}
\title{String Left Trim}
\usage{
s_ltrim(s, t = " \\t\\n\\r\\f\\v")
}
\arguments{
\item{s}{String to be trimmed}

\item{t}{Trimming elements}
}
\value{
Updated String s, left trimmed
}
\description{
String Left Trim
}
