% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s_md5}
\alias{s_md5}
\title{Compute md5 of single element}
\usage{
s_md5(str)
}
\arguments{
\item{str}{Single String Element}
}
\value{
single MD5 (times = 1)
}
\description{
Compute md5 of single element
}
