% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s_rtrim}
\alias{s_rtrim}
\title{String Right Trim}
\usage{
s_rtrim(s, t = " \\t\\n\\r\\f\\v")
}
\arguments{
\item{s}{String to be trimmed}

\item{t}{Trimming elements}
}
\value{
Updated String s, right trimmed
}
\description{
String Right Trim
}
