% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s_str_repeat}
\alias{s_str_repeat}
\title{Repeat a String s}
\usage{
s_str_repeat(s, times)
}
\arguments{
\item{String}{to 'repeat'}

\item{Integer}{'times' to 'repeat'}
}
\value{
updated String repeated 'times'
}
\description{
Repeat a String s
}
