% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s_tolower}
\alias{s_tolower}
\title{String to lower case}
\usage{
s_tolower(s, locale = "en_US.UTF-8")
}
\arguments{
\item{s}{String to be transformed}
}
\value{
Updated String s, lower cased
}
\description{
String to lower case
}
