% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s_toupper}
\alias{s_toupper}
\title{String to upper case}
\usage{
s_toupper(s, locale = "en_US.UTF-8")
}
\arguments{
\item{s}{String to be transformed}
}
\value{
Updated String s, upper cased
}
\description{
String to upper case
}
