% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_trim}
\alias{cpp_trim}
\title{String  Trim}
\usage{
cpp_trim(str, side = "both", t = " \\t\\n\\r\\f\\v")
}
\arguments{
\item{t}{Trimming elements}

\item{s}{String to be trimmed}
}
\value{
Updated String s, trimmed (both left and right)
}
\description{
String  Trim
}
